+++ common/unit_test/Test_Common_Utils.cpp
@@ -27,11 +27,6 @@ using paired_int_types =
 
 // Basically the same fixtures, used for labeling tests
 template <typename T>
-struct TestConvertNegativeAxes : public ::testing::Test {
-  using value_type = T;
-};
-
-template <typename T>
 struct ContainerTypes : public ::testing::Test {
   static constexpr std::size_t rank = 5;
   using value_type                  = T;
@@ -55,296 +50,6 @@ struct TestConvertBaseTypes : public ::testing::Test {
   using value_type1 = typename T::first_type;
   using value_type2 = typename T::second_type;
 };
-
-// Tests for convert_negative_axes over ND views
-template <typename IntType>
-void test_convert_negative_axes_1d() {
-  constexpr std::size_t DIM = 1, Rank = 1;
-  using array_type = std::array<IntType, DIM>;
-
-  auto converted_axes_0 =
-      KokkosFFT::Impl::convert_negative_axes(array_type{0}, Rank);
-
-  array_type ref_converted_axes_0 = {0};
-  EXPECT_EQ(converted_axes_0, ref_converted_axes_0);
-
-  // Check if errors are correctly raised against invalid axis
-  // axis must be in [-1, 1)
-  EXPECT_THROW(
-      ({ KokkosFFT::Impl::convert_negative_axes(array_type{1}, Rank); }),
-      std::runtime_error);
-
-  if constexpr (std::is_signed_v<IntType>) {
-    array_type converted_axes_minus1 =
-        KokkosFFT::Impl::convert_negative_axes(array_type{-1}, Rank);
-    array_type ref_converted_axes_minus1 = {0};
-    EXPECT_EQ(converted_axes_minus1, ref_converted_axes_minus1);
-
-    EXPECT_THROW(
-        ({ KokkosFFT::Impl::convert_negative_axes(array_type{-2}, Rank); }),
-        std::runtime_error);
-  }
-}
-
-template <typename IntType>
-void test_convert_negative_axes_2d() {
-  constexpr std::size_t Rank = 2;
-  using array_1d_type        = std::array<IntType, 1>;
-  using array_2d_type        = std::array<IntType, 2>;
-
-  auto converted_axes_0 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{0}, Rank);
-  auto converted_axes_1 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{1}, Rank);
-  auto converted_axes_01 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{0, 1}, Rank);
-
-  array_1d_type ref_converted_axis_0  = {0};
-  array_1d_type ref_converted_axis_1  = {1};
-  array_2d_type ref_converted_axes_01 = {0, 1};
-
-  EXPECT_EQ(converted_axes_0, ref_converted_axis_0);
-  EXPECT_EQ(converted_axes_1, ref_converted_axis_1);
-  EXPECT_EQ(converted_axes_01, ref_converted_axes_01);
-
-  // Check if errors are correctly raised against invalid axis
-  // axis must be in [-2, 2)
-  EXPECT_THROW(
-      ({ KokkosFFT::Impl::convert_negative_axes(array_1d_type{2}, Rank); }),
-      std::runtime_error);
-
-  if constexpr (std::is_signed_v<IntType>) {
-    array_1d_type ref_converted_axes_minus1  = {1};
-    array_2d_type ref_converted_axes_minus21 = {0, 1};
-    auto converted_axes_minus1 =
-        KokkosFFT::Impl::convert_negative_axes(array_1d_type{-1}, Rank);
-    auto converted_axes_minus21 =
-        KokkosFFT::Impl::convert_negative_axes(array_2d_type{-2, -1}, Rank);
-    EXPECT_EQ(converted_axes_minus1, ref_converted_axes_minus1);
-    EXPECT_EQ(converted_axes_minus21, ref_converted_axes_minus21);
-
-    EXPECT_THROW(
-        ({ KokkosFFT::Impl::convert_negative_axes(array_1d_type{-3}, Rank); }),
-        std::runtime_error);
-  }
-}
-
-template <typename IntType>
-void test_convert_negative_axes_3d() {
-  constexpr std::size_t Rank = 3;
-  using array_1d_type        = std::array<IntType, 1>;
-  using array_2d_type        = std::array<IntType, 2>;
-  using array_3d_type        = std::array<IntType, 3>;
-
-  auto converted_axes_0 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{0}, Rank);
-  auto converted_axes_1 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{1}, Rank);
-  auto converted_axes_2 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{2}, Rank);
-  auto converted_axes_01 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{0, 1}, Rank);
-  auto converted_axes_02 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{0, 2}, Rank);
-  auto converted_axes_12 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{1, 2}, Rank);
-  auto converted_axes_012 =
-      KokkosFFT::Impl::convert_negative_axes(array_3d_type{0, 1, 2}, Rank);
-
-  array_1d_type ref_converted_axes_0   = {0};
-  array_1d_type ref_converted_axes_1   = {1};
-  array_1d_type ref_converted_axes_2   = {2};
-  array_2d_type ref_converted_axes_01  = {0, 1};
-  array_2d_type ref_converted_axes_02  = {0, 2};
-  array_2d_type ref_converted_axes_12  = {1, 2};
-  array_3d_type ref_converted_axes_012 = {0, 1, 2};
-
-  EXPECT_EQ(converted_axes_0, ref_converted_axes_0);
-  EXPECT_EQ(converted_axes_1, ref_converted_axes_1);
-  EXPECT_EQ(converted_axes_2, ref_converted_axes_2);
-  EXPECT_EQ(converted_axes_01, ref_converted_axes_01);
-  EXPECT_EQ(converted_axes_02, ref_converted_axes_02);
-  EXPECT_EQ(converted_axes_12, ref_converted_axes_12);
-  EXPECT_EQ(converted_axes_012, ref_converted_axes_012);
-
-  // Check if errors are correctly raised against invalid axis
-  // axis must be in [-3, 3)
-  EXPECT_THROW(
-      ({ KokkosFFT::Impl::convert_negative_axes(array_1d_type{3}, Rank); }),
-      std::runtime_error);
-
-  EXPECT_THROW(
-      ({
-        KokkosFFT::Impl::convert_negative_axes(array_2d_type{3, 1}, Rank);
-      }),
-      std::runtime_error);
-
-  EXPECT_THROW(
-      ({
-        KokkosFFT::Impl::convert_negative_axes(array_3d_type{0, 1, 3}, Rank);
-      }),
-      std::runtime_error);
-
-  if constexpr (std::is_signed_v<IntType>) {
-    auto converted_axes_minus1 =
-        KokkosFFT::Impl::convert_negative_axes(array_1d_type{-1}, Rank);
-    auto converted_axes_minus2 =
-        KokkosFFT::Impl::convert_negative_axes(array_1d_type{-2}, Rank);
-    auto converted_axes_minus3 =
-        KokkosFFT::Impl::convert_negative_axes(array_1d_type{-3}, Rank);
-
-    auto converted_axes_minus21 =
-        KokkosFFT::Impl::convert_negative_axes(array_2d_type{-2, -1}, Rank);
-    auto converted_axes_minus321 =
-        KokkosFFT::Impl::convert_negative_axes(array_3d_type{-3, -2, -1}, Rank);
-
-    array_1d_type ref_converted_axes_minus1   = {2},
-                  ref_converted_axes_minus2   = {1},
-                  ref_converted_axes_minus3   = {0};
-    array_2d_type ref_converted_axes_minus21  = {1, 2};
-    array_3d_type ref_converted_axes_minus321 = {0, 1, 2};
-
-    EXPECT_EQ(converted_axes_minus1, ref_converted_axes_minus1);
-    EXPECT_EQ(converted_axes_minus2, ref_converted_axes_minus2);
-    EXPECT_EQ(converted_axes_minus3, ref_converted_axes_minus3);
-    EXPECT_EQ(converted_axes_minus21, ref_converted_axes_minus21);
-    EXPECT_EQ(converted_axes_minus321, ref_converted_axes_minus321);
-
-    EXPECT_THROW(
-        ({ KokkosFFT::Impl::convert_negative_axes(array_1d_type{-4}, Rank); }),
-        std::runtime_error);
-  }
-}
-
-template <typename IntType>
-void test_convert_negative_axes_4d() {
-  constexpr std::size_t Rank = 4;
-  using array_1d_type        = std::array<IntType, 1>;
-  using array_2d_type        = std::array<IntType, 2>;
-  using array_3d_type        = std::array<IntType, 3>;
-  using array_4d_type        = std::array<IntType, 4>;
-
-  auto converted_axes_0 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{0}, Rank);
-  auto converted_axes_1 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{1}, Rank);
-  auto converted_axes_2 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{2}, Rank);
-  auto converted_axes_3 =
-      KokkosFFT::Impl::convert_negative_axes(array_1d_type{3}, Rank);
-  auto converted_axes_01 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{0, 1}, Rank);
-  auto converted_axes_02 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{0, 2}, Rank);
-  auto converted_axes_03 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{0, 3}, Rank);
-  auto converted_axes_12 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{1, 2}, Rank);
-  auto converted_axes_13 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{1, 3}, Rank);
-  auto converted_axes_23 =
-      KokkosFFT::Impl::convert_negative_axes(array_2d_type{2, 3}, Rank);
-  auto converted_axes_012 =
-      KokkosFFT::Impl::convert_negative_axes(array_3d_type{0, 1, 2}, Rank);
-  auto converted_axes_013 =
-      KokkosFFT::Impl::convert_negative_axes(array_3d_type{0, 1, 3}, Rank);
-  auto converted_axes_023 =
-      KokkosFFT::Impl::convert_negative_axes(array_3d_type{0, 2, 3}, Rank);
-  auto converted_axes_123 =
-      KokkosFFT::Impl::convert_negative_axes(array_3d_type{1, 2, 3}, Rank);
-  auto converted_axes_0123 =
-      KokkosFFT::Impl::convert_negative_axes(array_4d_type{0, 1, 2, 3}, Rank);
-
-  array_1d_type ref_converted_axes_0    = {0};
-  array_1d_type ref_converted_axes_1    = {1};
-  array_1d_type ref_converted_axes_2    = {2};
-  array_1d_type ref_converted_axes_3    = {3};
-  array_2d_type ref_converted_axes_01   = {0, 1};
-  array_2d_type ref_converted_axes_02   = {0, 2};
-  array_2d_type ref_converted_axes_03   = {0, 3};
-  array_2d_type ref_converted_axes_12   = {1, 2};
-  array_2d_type ref_converted_axes_13   = {1, 3};
-  array_2d_type ref_converted_axes_23   = {2, 3};
-  array_3d_type ref_converted_axes_012  = {0, 1, 2};
-  array_3d_type ref_converted_axes_013  = {0, 1, 3};
-  array_3d_type ref_converted_axes_023  = {0, 2, 3};
-  array_3d_type ref_converted_axes_123  = {1, 2, 3};
-  array_4d_type ref_converted_axes_0123 = {0, 1, 2, 3};
-
-  EXPECT_EQ(converted_axes_0, ref_converted_axes_0);
-  EXPECT_EQ(converted_axes_1, ref_converted_axes_1);
-  EXPECT_EQ(converted_axes_2, ref_converted_axes_2);
-  EXPECT_EQ(converted_axes_3, ref_converted_axes_3);
-  EXPECT_EQ(converted_axes_01, ref_converted_axes_01);
-  EXPECT_EQ(converted_axes_02, ref_converted_axes_02);
-  EXPECT_EQ(converted_axes_03, ref_converted_axes_03);
-  EXPECT_EQ(converted_axes_12, ref_converted_axes_12);
-  EXPECT_EQ(converted_axes_13, ref_converted_axes_13);
-  EXPECT_EQ(converted_axes_23, ref_converted_axes_23);
-  EXPECT_EQ(converted_axes_012, ref_converted_axes_012);
-  EXPECT_EQ(converted_axes_013, ref_converted_axes_013);
-  EXPECT_EQ(converted_axes_023, ref_converted_axes_023);
-  EXPECT_EQ(converted_axes_123, ref_converted_axes_123);
-  EXPECT_EQ(converted_axes_0123, ref_converted_axes_0123);
-
-  // Check if errors are correctly raised against invalid axis
-  // axis must be in [-4, 4)
-  EXPECT_THROW(
-      ({ KokkosFFT::Impl::convert_negative_axes(array_1d_type{4}, Rank); }),
-      std::runtime_error);
-
-  EXPECT_THROW(
-      ({
-        KokkosFFT::Impl::convert_negative_axes(array_2d_type{4, 1}, Rank);
-      }),
-      std::runtime_error);
-
-  EXPECT_THROW(
-      ({
-        KokkosFFT::Impl::convert_negative_axes(array_3d_type{0, 1, 4}, Rank);
-      }),
-      std::runtime_error);
-
-  EXPECT_THROW(
-      ({
-        KokkosFFT::Impl::convert_negative_axes(array_4d_type{0, 1, 2, 4}, Rank);
-      }),
-      std::runtime_error);
-
-  if constexpr (std::is_signed_v<IntType>) {
-    auto converted_axes_minus1 =
-        KokkosFFT::Impl::convert_negative_axes(array_1d_type{-1}, Rank);
-    auto converted_axes_minus2 =
-        KokkosFFT::Impl::convert_negative_axes(array_1d_type{-2}, Rank);
-    auto converted_axes_minus3 =
-        KokkosFFT::Impl::convert_negative_axes(array_1d_type{-3}, Rank);
-
-    auto converted_axes_minus21 =
-        KokkosFFT::Impl::convert_negative_axes(array_2d_type{-2, -1}, Rank);
-    auto converted_axes_minus321 =
-        KokkosFFT::Impl::convert_negative_axes(array_3d_type{-3, -2, -1}, Rank);
-    auto converted_axes_minus4321 = KokkosFFT::Impl::convert_negative_axes(
-        array_4d_type{-4, -3, -2, -1}, Rank);
-
-    array_1d_type ref_converted_axis_minus1    = {3},
-                  ref_converted_axis_minus2    = {2},
-                  ref_converted_axis_minus3    = {1};
-    array_2d_type ref_converted_axes_minus21   = {2, 3};
-    array_3d_type ref_converted_axes_minus321  = {1, 2, 3};
-    array_4d_type ref_converted_axes_minus4321 = {0, 1, 2, 3};
-
-    EXPECT_EQ(converted_axes_minus1, ref_converted_axis_minus1);
-    EXPECT_EQ(converted_axes_minus2, ref_converted_axis_minus2);
-    EXPECT_EQ(converted_axes_minus3, ref_converted_axis_minus3);
-    EXPECT_EQ(converted_axes_minus21, ref_converted_axes_minus21);
-    EXPECT_EQ(converted_axes_minus321, ref_converted_axes_minus321);
-    EXPECT_EQ(converted_axes_minus4321, ref_converted_axes_minus4321);
-
-    EXPECT_THROW(
-        ({ KokkosFFT::Impl::convert_negative_axes(array_1d_type{-5}, Rank); }),
-        std::runtime_error);
-  }
-}
 
 template <typename ContainerType>
 void test_reversed() {
@@ -597,32 +302,10 @@ void test_convert_base_int_type() {
 }
 }  // namespace
 
-TYPED_TEST_SUITE(TestConvertNegativeAxes, base_int_types);
 TYPED_TEST_SUITE(ContainerTypes, base_int_types);
 TYPED_TEST_SUITE(PairedScalarTypes, paired_scalar_types);
 TYPED_TEST_SUITE(TestIndexSequence, base_int_types);
 TYPED_TEST_SUITE(TestConvertBaseTypes, paired_int_types);
-
-// Tests for 1D - 4D View
-TYPED_TEST(TestConvertNegativeAxes, 1DView) {
-  using value_type = typename TestFixture::value_type;
-  test_convert_negative_axes_1d<value_type>();
-}
-
-TYPED_TEST(TestConvertNegativeAxes, 2DView) {
-  using value_type = typename TestFixture::value_type;
-  test_convert_negative_axes_2d<value_type>();
-}
-
-TYPED_TEST(TestConvertNegativeAxes, 3DView) {
-  using value_type = typename TestFixture::value_type;
-  test_convert_negative_axes_3d<value_type>();
-}
-
-TYPED_TEST(TestConvertNegativeAxes, 4DView) {
-  using value_type = typename TestFixture::value_type;
-  test_convert_negative_axes_4d<value_type>();
-}
 
 TYPED_TEST(ContainerTypes, test_total_size_of_vector) {
   using container_type = typename TestFixture::vector_type;
